Refactor extracting TransferOptions from Settings
